prune unnecessary dependencies, add all extra, add test pipelines for core and all dependency extras, delete unused cache_helper, update deprecated datetime utcnow method, fix HF target file name, move HF test file to target folder
