+++ .github/workflows/build_and_test.yml
@@ -25,6 +25,7 @@ jobs:
         os: [ubuntu-latest]
         python: ["3.10", "3.11", "3.12"]
         package_name: ["pyrit"]
+        package_extras: ["dev", "all"]
     runs-on: ${{ matrix.os }}
     # EnricoMi/publish-unit-test-result-action@v2 requires the following permissions
     permissions:
@@ -40,7 +41,7 @@ jobs:
       - name: Install setuptools and pip
         run: pip install --upgrade setuptools pip
       - name: Install PyRIT with pip
-        run: pip install .[dev,gcg]
+        run: pip install .[${{ matrix.package_extras }}]
       - name: Run unit tests with code coverage
         run: make test-cov-xml
       - name: Publish Pytest Results
