+++ pyproject.toml
@@ -50,13 +50,9 @@ dependencies = [
     "duckdb==0.10.0",
     "duckdb-engine==0.11.2",
     "httpx[http2]>=0.27.2",
-    "jsonpickle>=3.0.4",
     "jupyter>=1.0.0",
     "ipykernel>=6.29.4",
-    "logzero>=1.7.0",
     "numpy>=1.26.4",
-    "onnxruntime>=1.14.1",
-    "onnx>=1.16.0",
     "openai>=1.52.2",
     "pillow>=10.3.0",
     "pydantic>=2.7.1",
@@ -70,15 +66,14 @@ dependencies = [
     "sqlalchemy-stubs==0.4",
     "termcolor>=2.4.0",
     "tenacity>=8.2.3",
-    "tokenizers>=0.19.1",
     "tqdm>=4.66.2",
     "transformers>=4.40.0",
     "treelib>=1.7.0",
     "types-requests>=2.31.0.20240406",
-    "torch==2.4.1"
 ]
 
 [project.optional-dependencies]
+# always make sure the individual ones are in sync with the all group
 dev = [
     "black>=24.4.0",
     "flake8>=7.0.0",
@@ -96,16 +91,38 @@ dev = [
     "types-PyYAML>=6.0.12.9",
 ]
 torch = [
-    "torch>=2.3.0"
+    "torch>=2.3.0",
 ]
 gcg = [
-    "numpy<2",
+    "accelerate==0.34.2",
+    "azureml-mlflow==1.57.0",
     "mlflow==2.16.2",
     "ml-collections==0.1.1",
-    "azureml-mlflow==1.57.0",
-    "accelerate==0.34.2",
-    "transformers==4.45.1",
+    "numpy<2",
     "sentencepiece==0.2.0",
+]
+all = [
+    "accelerate==0.34.2",
+    "azureml-mlflow==1.57.0",
+    "black>=24.4.0",
+    "flake8>=7.0.0",
+    "flake8-copyright>=0.2.4",
+    "jupyter-book>=1.0.2",
+    "jupytext>=1.16.1",
+    "mlflow==2.16.2",
+    "ml-collections==0.1.1",
+    "mypy>=1.9.0",
+    "mock-alchemy>=0.2.6",
+    "numpy<2",
+    "pre-commit>=3.3.3",
+    "pytest>=7.3.1",
+    "pytest-asyncio>=0.23.5",
+    "pytest-cov>=4.0.0",
+    "respx>=0.20.2",
+    "semantic-kernel==0.9.4b1",
+    "sentencepiece==0.2.0",
+    "torch>=2.3.0",
+    "types-PyYAML>=6.0.12.9",
 ]
 
 [tool.pytest.ini_options]
