+++ pyrit/prompt_target/__init__.py
@@ -17,7 +17,7 @@ from pyrit.prompt_target.crucible_target import CrucibleTarget
 from pyrit.prompt_target.gandalf_target import GandalfTarget, GandalfLevel
 from pyrit.prompt_target.http_target.http_target import HTTPTarget
 from pyrit.prompt_target.hugging_face.hugging_face_chat_target import HuggingFaceChatTarget
-from pyrit.prompt_target.hugging_face.hugging_face_enpoint_target import HuggingFaceEndpointTarget
+from pyrit.prompt_target.hugging_face.hugging_face_endpoint_target import HuggingFaceEndpointTarget
 from pyrit.prompt_target.ollama_chat_target import OllamaChatTarget
 from pyrit.prompt_target.prompt_shield_target import PromptShieldTarget
 from pyrit.prompt_target.text_target import TextTarget
